Se corrige clase Simulador se incluyen los metodos solicitados

+++ Algoritmos.h
@@ -15,7 +15,7 @@ using namespace std;
 class Algoritmos
 {
 public:
-    static void quicksortEquipo (Equipo** equipos, double lo, double high ); // equipos apuntaria a un vector de puntesros
+    static void quicksortEquipo (Equipo** equipos, int lo, int high ); // equipos apuntaria a un vector de puntesros
     static void ordenarEquiposPorPrioridadDesc(Equipo** equipo); // Equipo apuntaria a un vector de punteros
     static int busquedaBinariaId(string** id, string idBuscado); // id apunta a un vector de punteros
 };
